% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbpositpack-package.R
\docType{package}
\name{rbpositpack-package}
\alias{rbpositpack}
\alias{rbpositpack-package}
\title{rbpositpack: Count the installed packages}
\description{
Was ceated during PositConf 2023. Has no real value.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/greenhorn82/rbpositpack}
  \item \url{https://greenhorn82.github.io/rbpositpack/}
  \item Report bugs at \url{https://github.com/greenhorn82/rbpositpack/issues}
}

}
\author{
\strong{Maintainer}: Robert Busching \email{busching@uni-potsdam.de} (\href{https://orcid.org/0000-0001-7522-6053}{ORCID})

}
\keyword{internal}
